WIP add server class for api
