+++ public_html/server/classes/appmonitor-server.class.php
@@ -134,6 +134,16 @@ class appmonitorserver {
         return $this->oLang = new lang($this->_aCfg['lang']);
     }
     
+    /**
+     * get a flat array with all application ids and website + url
+     * as subkeys
+     * @return array
+     */
+    public function getAppIds() {
+        $this->_getClientData(true);
+        return array_keys($this->_data);
+    }
+
     public function getConfigVars() {
         return $this->_aCfg;
     }
@@ -179,6 +189,17 @@ class appmonitorserver {
             'notifications' => $this->_aCfg['notifications']
         ));
     }
+    /**
+     * load monitoring data ... if not done yet; used in gui and api
+     * @return boolean
+     */
+    public function loadClientData(){
+        if (!count($this->_data)) {
+            $this->_getClientData();
+        }
+        return true;
+    }
+    
 
     /**
      * save the current config
@@ -295,6 +316,24 @@ class appmonitorserver {
             default:
                 break;
         }
+    }
+
+    /**
+     * get a flat array of tags sent from all clients
+     * @return array
+     */
+    protected function _getClientTags(){
+        $aTags=array();
+        foreach ($this->_data as $aEntries) {
+            if (isset($aEntries['meta']['tags'])){
+                foreach($aEntries['meta']['tags'] as $sTag){
+                    $aTags[]=$sTag;
+                }
+            }
+        }
+        sort($aTags);
+        $aTags = array_unique($aTags);
+        return $aTags;
     }
 
     /**
@@ -725,77 +764,6 @@ class appmonitorserver {
         );
     }
 
-    /**
-     * get array with application data 
-     * 
-     * @param string  $sKey           filter key i.e. "meta"; default false (all)
-     * @param string  $sFilterAppId   filter by app id; default false (all)
-     * @return array
-     */
-    protected function _apiGetAppData($sKey=false, $sFilterAppId=false) {
-        $this->_getClientData(true); // get data; true = use cache
-        $aReturn=array();
-        // echo 'ALL client data<pre>'.print_r($this->_data, 1).'</pre>';
-        // echo '<br>$sKey = '.$sKey.'<br>';
-        if ($sFilterAppId && !isset($this->_data[$sFilterAppId])){
-            $aReturn=['error'=>'App id was not found', 'http' => '404'];
-        } else {
-            foreach($this->_data as $sAppId=>$aData){
-                if($sAppId===$sFilterAppId){
-                    $aReturn=$sKey 
-                        ? $aData[$sKey] 
-                        : $aData;
-                } else {
-                    $aReturn[$sAppId]=$sKey 
-                        ? $aData[$sKey] 
-                        : $aData;
-                }
-            }
-        }
-        return $aReturn;
-    }
-
-    /**
-     * get a flat array with all application ids
-     * @return array
-     */
-    public function apiGetAppIds() {
-        $this->_getClientData(true);
-        return array_keys($this->_data);
-    }
-    
-    /**
-     * get an array of all client data; optional filtere by given app id 
-     * @param string  $sFilterAppId   filter by app id; default false (all)
-     * @return array
-     */
-    public function apiGetAppAllData($sFilterAppId=false) {
-        return $this->_apiGetAppData(false, $sFilterAppId);
-    }
-    /**
-     * get an array of all client checks; optional filtered by given app id 
-     * @param string  $sFilterAppId   filter by app id; default false (all)
-     * @return type
-     */
-    public function apiGetAppChecks($sFilterAppId=false) {
-        return $this->_apiGetAppData('checks',$sFilterAppId);
-    }
-    /**
-     * get an array of all client metadata; optional filtered by given app id 
-     * @param string  $sFilterAppId   filter by app id; default false (all)
-     * @return type
-     */
-    public function apiGetAppMeta($sFilterAppId=false) {
-        return $this->_apiGetAppData('meta',$sFilterAppId);
-    }
-    /**
-     * get an array of all client metadata; optional filtered by given app id 
-     * @param string  $sFilterAppId   filter by app id; default false (all)
-     * @return type
-     */
-    public function apiGetTroubleItems($sFilterAppId=false) {
-        return $this->_apiGetAppData('meta',$sFilterAppId);
-    }
     /**
      * set flag for logging to standard output
      */
